requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.3
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.3.1
huggingface_hub==0.24.6
mcp
ollama
